Reorganize dashboard JS code
